Add files via upload
Levels testify as an important game element for
Procedural Content Generation. Recently, machine learning techniques
have been used for the automatic content generation
of games. However, there is a need for a corpus for training
purposes. Considering this, we provide a corpus for the famous
angry birds game. Our corpus contains 200 levels in total, with
100 levels from the original Angry Birds game and 100 from the
Baseline level generator. These levels are encoded using a tilebased
approach and are freely available for applying machine
learning techniques. Further to this, our next step would be to
mimic the style of these levels in the forthcoming angry birds
level generation competition.